feat: action page (#2006)
