--- apps/renderer/src/pages/settings/(settings)/actions.tsx
@@ -1,24 +1,0 @@
-import { UserRole } from "@follow/constants"
-
-import { ActionSetting } from "~/modules/settings/tabs/actions"
-import { SettingsTitle } from "~/modules/settings/title"
-import { defineSettingPageData, DisableWhy } from "~/modules/settings/utils"
-
-const iconName = "i-mgc-magic-2-cute-re"
-const priority = 1020
-
-export const loader = defineSettingPageData({
-  icon: iconName,
-  name: "titles.actions",
-  priority,
-  disableIf: (ctx) => [ctx.role === UserRole.Trial, DisableWhy.NotActivation],
-})
-
-export function Component() {
-  return (
-    <>
-      <SettingsTitle />
-      <ActionSetting />
-    </>
-  )
-}
