+++ apps/renderer/src/queries/actions.ts
@@ -1,10 +1,12 @@
-import { apiClient } from "~/lib/api-fetch"
+import { useAuthQuery } from "~/hooks/common/useBizQuery"
 import { defineQuery } from "~/lib/defineQuery"
+import { actionActions } from "~/store/action"
 
 export const action = {
-  getAll: () =>
-    defineQuery(["actions"], async () => {
-      const res = await apiClient.actions.$get()
-      return res.data
-    }),
+  getAll: () => defineQuery(["actions"], () => actionActions.fillRemoteActions()),
 }
+
+export const useActionsQuery = () =>
+  useAuthQuery(action.getAll(), {
+    staleTime: Infinity,
+  })
