+++ apps/renderer/src/store/action/types.ts
@@ -1,0 +1,6 @@
+import type { ActionModel } from "@follow/models/types"
+
+export interface ActionState {
+  actions: ActionModel[]
+  isDirty: boolean
+}
