+++ changelog/next.md
@@ -5,6 +5,7 @@
 - New `Move to Category` operation in feed subscription context menu
 - New `Expand long social media` setting to automatically expand social media entries containing long text.
 - New `Back Top` button and read progress indicator in entry content
+- Independent Action page
 
 ## Improvements
