+++ locales/app/en.json
@@ -373,6 +373,7 @@
   "trending.user": "Trending Users",
   "user_button.account": "Account",
   "user_button.achievement": "Achievements",
+  "user_button.actions": "Actions",
   "user_button.download_desktop_app": "Download Desktop app",
   "user_button.log_out": "Log out",
   "user_button.power": "Power",
@@ -385,6 +386,7 @@
   "user_profile.share": "Share",
   "user_profile.toggle_item_style": "Toggle Item Style",
   "words.achievement": "Achievements",
+  "words.actions": "Actions",
   "words.add": "Add",
   "words.boost": "Boost",
   "words.browser": "Browser",
