+++ locales/settings/en.json
@@ -7,6 +7,7 @@
   "about.socialMedia": "Social Media",
   "actions.actionName": "Action {{number}}",
   "actions.action_card.add": "Add",
+  "actions.action_card.add_action": "Add Action",
   "actions.action_card.all": "All",
   "actions.action_card.and": "And",
   "actions.action_card.block": "Block",
